+++ ecmwf_models/download.py
@@ -22,12 +22,9 @@
 # SOFTWARE.
 
 '''
-Module for Downloading ECMWF ERA Interim and ERA5 data in grib and netcdf format.
+General functions to reorganise downloaded ERA file
 '''
 
-from ecmwfapi import ECMWFDataServer
-import argparse
-import sys
 import warnings
 try:
     import pygrib
@@ -35,91 +32,12 @@ except ImportError:
     warnings.warn("pygrib has not been imported")
 
 import os
-from datetime import datetime, timedelta
+from datetime import datetime
 import xarray as xr
 import pandas as pd
 from datedown.fname_creator import create_dt_fpath
-import shutil
 import numpy as np
 
-
-def download_era(start, end, parameters, target, product='ERA-Interim', format='grib1',
-                 grid_size=None, timesteps=[0, 6, 12, 18], landmask = True):
-    """
-    Download era 5 data
-
-    Parameters
-    ----------
-    start : date
-        start date
-    end : date
-        end date
-    parameters : list
-        parameter ids, see wiki
-    target : str
-        path at which to save the downloaded grib file
-    product : str, optional
-        Name of the model, "ERA-interim" (default) or "ERA5"
-    format: str, optional
-        format of the downloaded data, netcdf or grib1 (default)
-    grid_size: [float,float], optional
-        size of the grid in form (lon, lat), which the data is resampled to
-        If None is passed the minimum grid for the accoring product is chosen
-    timesteps: list
-        list of times for which data is downloaded
-    landmask: bool
-        If True, also download the land/sea mask
-    """
-    server = ECMWFDataServer()
-    param_strings = []
-
-    if product == 'ERA-Interim':
-        dataset = 'interim'
-        dataclass = 'ei'
-    elif product == 'ERA5':
-        dataset = 'era5'
-        dataclass = 'ea'
-    else:
-        raise ValueError('Unknown ECMWF product. Use "ecmwf_download -h" too show supported data sets')
-
-    if landmask and 172 not in parameters:
-        parameters.append(172)
-
-    for parameter in parameters:
-        param_strings.append("%d.128" % parameter)
-
-    timestep_strings = []
-    for timestep in timesteps:
-        timestep_strings.append("%02d" % timestep)
-
-    param_string = '/'.join(param_strings)
-    timestep_string = '/'.join(timestep_strings)
-    date_string = "%s/to/%s" % (start.strftime("%Y-%m-%d"),
-                                end.strftime("%Y-%m-%d"))
-
-    grid_size = "%f/%f" % (grid_size[0], grid_size[1]) if grid_size else None
-
-    # ATTENTION: When downloading netcdf files steps and times must not overlap!!
-    # see: https://software.ecmwf.int/wiki/display/CKB/What+to+do+with+ECCODES+ERROR+%3A+Try+using+the+-T+option
-    dl_params = {"class": dataclass, "dataset": dataset, "expver": "1", "stream": "oper", "type": "an", "levtype": "sfc",
-                 "param": param_string, "date": date_string, "time": timestep_string, "step": "0", "grid": grid_size,
-                 "format": format, "target": target}
-
-    if not grid_size:
-        if format == 'netcdf':
-            if product == 'ERA5':
-                grid_size = "%f/%f" % (0.3,0.3)
-            else:
-                grid_size = "%f/%f" % (0.75,0.75)
-            dl_params['grid'] = grid_size
-        else:
-            del dl_params['grid']
-    else:
-        if (any(size < 0.75 for size in grid_size) and product == 'ERA-Interim') or \
-           (any(size < 0.3 for size in grid_size) and product == 'ERA5'):
-            raise Warning('Custom grid smaller than original ERA data. See https://software.ecmwf.int/wiki/display/CKB/Does+downloading+data+at+higher+resolution+improve+the+output')
-
-    server.retrieve(dl_params)
 
 
 
@@ -214,68 +132,7 @@ def save_gribs_from_grib(input_grib, output_path, product_name,
     grib_in.close()
 
 
-def download_and_move(parameters, startdate, enddate, product, format,
-                      target_path, keep_original=False, grid_size=None, timesteps=[0, 6, 12, 18]):
-    """
-    Downloads the data from the ECMWF servers and moves them to the target path.
-    This is done in 30 day increments between start and end date to be efficient with the MARS system.
-    See the recommendation for doing it this way in
-    https://software.ecmwf.int/wiki/display/WEBAPI/ERA-Interim+daily+retrieval+efficiency
 
-    The files are then extracted into separate grib files per parameter and stored
-    in yearly folders under the target_path.
-
-    Parameters
-    ----------
-    parameters : list
-        parameter ids
-    startdate: datetime
-        first date to download
-    enddate: datetime
-        last date to download
-    product: str
-        Name of the dataset to download (eg. ERA5, ERA-Interim)
-    format: str
-        format of the dataset to download (eg netcdf, grib)
-    target_path: string
-        path to which to copy the extracted parameter grib files
-    keep_original: bool
-        keep the original downloaded data
-    grid_size: list
-        [lon, lat] extent of the grid (regular for netcdf, at lat=0 for grib)
-    timesteps: list, optional
-        list of timesteps to download
-    """
-    td = timedelta(days=30)
-    current_start = startdate
-
-    if format not in ['netcdf', 'grib1']:
-        raise ValueError("Choose 'grib1' or 'netcdf' as format")
-    fextension = '.grb' if format == 'grib1' else '.nc'
-
-
-    while current_start <= enddate:
-        current_end = current_start + td
-        if current_end >= enddate:
-            current_end = enddate
-
-        fname = current_start.strftime("%Y%m%d_")
-        fname = current_end.strftime(fname + "%Y%m%d" + fextension)
-        downloaded_data_path = os.path.join(target_path, 'temp_downloaded')
-        if not os.path.exists(downloaded_data_path):
-            os.mkdir(downloaded_data_path)
-        downloaded_data = os.path.join(downloaded_data_path, fname)
-
-        download_era(current_start, current_end, parameters, downloaded_data,
-                     product, format, grid_size, timesteps=timesteps)
-
-        if format == 'netcdf':
-            save_ncs_from_nc(downloaded_data, target_path, product)
-        else:
-            save_gribs_from_grib(downloaded_data, target_path, product)
-        if not keep_original:
-            shutil.rmtree(downloaded_data_path)
-        current_start = current_end + timedelta(days=1)
 
 
 def mkdate(datestring):
@@ -292,65 +149,3 @@ def mkdate(datestring):
         return datetime.strptime(datestring, '%Y-%m-%dT%H:%M')
 
 
-def parse_args(args):
-    """
-    Parse command line parameters for recursive download
-
-    :param args: command line parameters as list of strings
-    :return: command line parameters as :obj:`argparse.Namespace`
-    """
-    parser = argparse.ArgumentParser(
-        description="Download ECMWF reanalysis model data")
-    parser.add_argument("localroot",
-                        help='Root of local filesystem where the data is stored.')
-    parser.add_argument("parameters", metavar="parameters", type=int,
-                        nargs="+",
-                        help=("Parameters to convert in numerical format. e.g."
-                              "39 40 41 42 for Volumetric soil water layers 1 to 4."
-                              "A list of possible parameters is available at http://apps.ecmwf.int/codes/grib/param-db "
-                              "or by using the 'View MARS request' option in the web based ordering system."))
-    parser.add_argument("-s", "--start", type=mkdate,
-                        help=("Startdate. Either in format YYYY-MM-DD or YYYY-MM-DDTHH:MM."
-                              "If no data is found there then the first available date of the product is used."))
-    parser.add_argument("-e", "--end", type=mkdate,
-                        help=("Enddate. Either in format YYYY-MM-DD or YYYY-MM-DDTHH:MM."
-                              "If not given then the current date is used."))
-    parser.add_argument("-p", "--product", type=str, default='ERA-Interim',
-                        help=("ECMWF product, ERA-Interim (default) or ERA5"))
-    parser.add_argument("-f", "--format", type=str, default='grib',
-                        help=("Downloaded data format, grib1 (default) or netcdf. "
-                              "Info on GRIB: https://software.ecmwf.int/wiki/display/CKB/What+are+GRIB+files"))
-    parser.add_argument("--grid_size", type=float, default=None, nargs='+',
-                        help=("lon lat, Size of the grid that the data is stored to. "
-                              "Must be set when downloading as 'netcdf'. "
-                              "Should be at least (and is by default) (0.75,0.75) for ERA-Interim "
-                              "and (0.3,0.3) for ERA5"))
-
-    args = parser.parse_args(args)
-
-    # set defaults that can not be handled by argparse
-    if args.start is None:
-        args.start = datetime(1979, 1, 1)
-    if args.end is None:
-        args.end = datetime.now()
-
-
-    print("Downloading {} data from {} to {} into folder {}".format(args.product,
-                                                                    args.start.isoformat(),
-                                                                    args.end.isoformat(),
-                                                                    args.localroot))
-    return args
-
-
-def main(args):
-    args = parse_args(args)
-
-    download_and_move(args.parameters, args.start, args.end, args.product,
-                      args.format, args.localroot, grid_size=args.grid_size)
-
-
-def run():
-    main(sys.argv[1:])
-
-if __name__ == '__main__':
-    run()